==> 01-dominance_analaysis.Rmd <==
# (PART) DOMINANCE ANALYSIS {-}
# Data preparation

```{r load_data_tube_tests, warning=FALSE, comments="", message=FALSE, results='hide'}
tube_test_data <- read_csv("data/tube_tests.csv") %>%
      mutate(mouse = substr(mouse, 1, 5)) %>%
      mutate(cage_id = paste0("C", cage_id)) %>%
      mutate(time = factor(case_when(
          treatment == "t1" ~ 1,
          treatment == "t2" ~ 2,
          treatment == "t3" ~ 3,
          treatment == "t4" ~ 4,
          treatment == "t5" ~ 5,
          treatment == "t6" ~ 6,
          TRUE ~ NA))) %>%
      arrange(time)
```

```{r previsualise_tube_test_data, warning=FALSE, comments="", message=FALSE}
tube_test_data %>%
  head() %>%
  tt() |> 
      style_tt(j = 1, color = "#6e0f0f", background = "#e39d9d") |> 
      style_tt(i = 0, color = "white", background = "#a52929")
```

# Model testing

We tested five models with varying degrees of complexity to estimate the probabilities of each animal to win a tube-test encounter against any of their cage mates.

```{r tube_test_model_testing, warning=FALSE, comments="", message=FALSE, results='hide'}
cages=c("C03","C04","C05","C06","C10","C11","C12","C13")

model_performance1 <- data.frame()
model_performance2 <- data.frame()
for(cage in cages){
    #Filter table by cage
    cage_data <- tube_test_data %>% filter(cage_id == cage)
    #Model with temporal autocorrelation (NULL if the model does not converge)
    M1 <- tryCatch({glmmTMB(win_loss_binomial~1+(1|mouse)+(1|treatment:mouse), data = cage_data, family = binomial)}, warning = function(w){NULL})
    M2 <- tryCatch({glmmTMB(win_loss_binomial~1+ar1(time+0|mouse)+(1|treatment:mouse), data = cage_data, family = binomial)}, warning = function(w){NULL})
    M3 <- tryCatch({glmmTMB(win_loss_binomial~1+ar1(time+0|mouse), data = cage_data, family = binomial)}, warning = function(w){NULL})
    M4 <- tryCatch({glmmTMB(win_loss_binomial~1+(1|mouse), data = cage_data, family = binomial)}, warning = function(w){NULL})
    M5 <- tryCatch({glmmTMB(win_loss_binomial~1, data = cage_data, family = binomial)}, warning = function(w){NULL})
    
    aic <- c(M1=if(is.null(M1)){NA}else{AIC(M1)},
             M2=if(is.null(M2)){NA}else{AIC(M2)},
             M3=if(is.null(M3)){NA}else{AIC(M3)},
             M4=if(is.null(M4)){NA}else{AIC(M4)},
             M5=if(is.null(M5)){NA}else{AIC(M5)}) %>% sort()
    
    rank <- names(aic)
             
    # Append predictions to table
    model_performance1 <- bind_rows(model_performance1,rank %>% as.data.frame() %>% t() %>% as.data.frame())
    model_performance2 <- bind_rows(model_performance2,aic %>% as.data.frame() %>% t() %>% as.data.frame())
}

```

## Model performance

```{r tube_test_model_testing_aic, warning=FALSE, comments="", message=FALSE}
colnames(model_performance2) <- c("M1","M2","M3","M4","M5")
rownames(model_performance2) <- cages
model_performance2 %>% 
  rownames_to_column(var="Cage") %>%
  tt() |> 
      style_tt(j = 1, color = "#6e0f0f", background = "#e39d9d") |> 
      style_tt(i = 0, color = "white", background = "#a52929")
```

## Model rank

```{r tube_test_model_testing_rank, warning=FALSE, comments="", message=FALSE}
colnames(model_performance1) <- c("1","2","3","4","5")
rownames(model_performance1) <- cages
model_performance1 %>% 
  rownames_to_column(var="Cage") %>%
  tt() |> 
      style_tt(j = 1, color = "#6e0f0f", background = "#e39d9d") |> 
      style_tt(i = 0, color = "white", background = "#a52929")
```
# Model prediction

```{r tube_test_modelling_prediction, warning=FALSE, comments="", message=FALSE, results='hide'}
cages=c("C03","C04","C05","C06","C10","C11","C12","C13")

dominance_predictions <- data.frame()
  
for(cage in cages){
    #Filter table by cage
    cage_data <- tube_test_data %>% filter(cage_id == cage)
    #Model with temporal autocorrelation (NULL if the model does not converge)
    M3 <- tryCatch({glmmTMB(win_loss_binomial~1+ar1(time+0|mouse), data = cage_data,family = binomial)}, warning = function(w){NULL})
    #Model without temporal autocorrelation
    M4 <- glmmTMB(win_loss_binomial~1+(1|mouse), data = cage_data,family = binomial)
    
    if(!is.null(M3)){
          cage_model <- data.frame(ggpredict(M3, type = "random", terms = c("mouse","time")), model="M3", cage=cage)
    }else{
          cage_model <- data.frame(ggpredict(M4, type = "random", terms = "mouse"), model="M4", cage=cage)
          cage_model <- bind_rows(cage_model %>% mutate(group = factor(1)),
            cage_model %>% mutate(group = factor(2)),
            cage_model %>% mutate(group = factor(3)),
            cage_model %>% mutate(group = factor(4)),
            cage_model %>% mutate(group = factor(5)),
            cage_model %>% mutate(group = factor(6)))}
    
    # Append predictions to table
    dominance_predictions <- bind_rows(dominance_predictions,cage_model)
}

dominance_predictions <- dominance_predictions %>%
      mutate(group = factor(case_when(
          group == 1 ~ "t1",
          group == 2 ~ "t2",
          group == 3 ~ "t3",
          group == 4 ~ "t4",
          group == 5 ~ "t5",
          group == 6 ~ "t6",
          TRUE ~ NA))) %>%
      rename(mouse=x, dominance=predicted, time=group)
```

## Dominance visualisation

```{r dominance_visualisation, warning=FALSE, comments="", message=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
dominance_predictions %>%
    mutate(time = as.numeric(substr(time, 2, 2))) %>%
    mutate(mouse_number = substr(mouse, 5, 5)) %>%
    ggplot(aes(x=time,y=dominance, group=mouse, color=mouse_number)) +
      geom_line(size=1) +
      scale_color_manual(values=c("#D5CFCF","#86b9d4","#a6897c","#2e2e33","#2e7287"))+
      scale_y_continuous(breaks = c(0,0.25,0.5,0.75,1)) +
      scale_x_continuous(breaks = c(1:6)) +
      facet_wrap(cage ~ ., ncol=4) + 
      theme_minimal()+
      theme(legend.position = "none")

```

## Hierarchy stability

```{r hierarchy_stability, warning=FALSE, comments="", message=FALSE}
cages=c("C03","C04","C05","C06","C10","C11","C12","C13")

hierarchy_stability <- c()
  
for(cage_code in cages){
    dominance_predictions_cage <- dominance_predictions %>% 
        filter(cage == cage_code) %>% 
        select(mouse,dominance,time) %>% 
        pivot_wider(names_from="time",values_from="dominance")%>% 
        select(-mouse)
    
    correlation_coefficients <- numeric(length = 5)
    for (i in 1:5) {
      correlation_coefficients[i] <- cor(dominance_predictions_cage[, i], dominance_predictions_cage[, i + 1], method = "spearman")
    }
    hierarchy_stability <- c(hierarchy_stability,mean(correlation_coefficients))
}

names(hierarchy_stability) <- cages

hierarchy_stability %>% 
    as.data.frame() %>% t() %>% as.data.frame() %>%
    tt(digits = 2) |> 
      style_tt(i = 0, color = "white", background = "#a52929")

mean(hierarchy_stability)
sd(hierarchy_stability)
```

```{r hierarchy_stability_plot, warning=FALSE, comments="", message=FALSE, results='hide'}
hierarchy_stability %>%
    as.data.frame() %>% t() %>% t() %>% as.data.frame() %>%
    rownames_to_column(var="cage") %>%
    rename(cage=1,stability=2) %>%
    ggplot(aes(y=fct_rev(cage),x=stability))+
      geom_col(fill="#999999")+
      theme_minimal()
```
### Between environmental treatments

```{r hierarchy_stability_treatments, warning=FALSE, comments="", message=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
hierarchy_stability %>% 
    as.data.frame() %>% t() %>% t() %>% as.data.frame() %>%
    rownames_to_column(var="cage") %>%
    rename(cage=1,stability=2) %>%
    mutate(treatment=c("variable","invariable","variable","variable","variable","variable","variable","invariable")) %>%
    group_by(treatment) %>%
    summarise(mean=mean(stability), sd=sd(stability)) %>%
    tt()
```

```{r hierarchy_stability_treatments_plot, warning=FALSE, comments="", message=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
hierarchy_stability %>% 
    as.data.frame() %>% t() %>% t() %>% as.data.frame() %>%
    rownames_to_column(var="cage") %>%
    rename(cage=1,stability=2) %>%
    mutate(treatment=c("variable","invariable","variable","variable","variable","variable","variable","invariable")) %>%
    ggplot(aes(x=treatment,y=stability,group=treatment))+
      geom_boxplot()
```

### Between sexes

```{r hierarchy_stability_sex, warning=FALSE, comments="", message=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
hierarchy_stability %>% 
    as.data.frame() %>% t() %>% t() %>% as.data.frame() %>%
    rownames_to_column(var="cage") %>%
    rename(cage=1,stability=2) %>%
    mutate(sex=c("male","male","male","male","female","female","female","female")) %>%
    group_by(sex) %>%
    summarise(mean=mean(stability), sd=sd(stability)) %>%
    tt()
```

```{r hierarchy_stability_sex_plot, warning=FALSE, comments="", message=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
hierarchy_stability %>% 
    as.data.frame() %>% t() %>% t() %>% as.data.frame() %>%
    rownames_to_column(var="cage") %>%
    rename(cage=1,stability=2) %>%
    mutate(sex=c("male","male","male","male","female","female","female","female")) %>%
    ggplot(aes(x=sex,y=stability,group=sex))+
      geom_boxplot()
```
